feat: properly calculates commission with a cliff

+++ database/migrations/2023_07_25_093354_create_cliffs_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('cliffs', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(Plan::class);
-            $table->integer('threshold_in_percent');
+            $table->float('threshold_in_percent');
             $table->timestamps();
         });
     }
